+++ app/src/main/java/com/example/fgapp/model/Model.kt
@@ -5,5 +5,5 @@ import java.util.concurrent.Future
 interface Model {
     fun connect(ip:String, port:Int): Future<Boolean>?
     fun disconnect()
-    fun changeVar(varName: String, value: Double)
+    fun changeVar(varName: String, value: Float)
 }
